+++ .env.development
@@ -1,3 +1,6 @@
+URL=https://app.outline.dev:3000
+
+SMTP_FROM_EMAIL=hello@example.com
 
 # Enable unsafe-inline in script-src CSP directive
 # Setting it to true allows React dev tools add-on in Firefox to successfully detect the project
