+++ app/actions/definitions/navigation.tsx
@@ -26,7 +26,6 @@ import SearchQuery from "~/models/SearchQuery";
 import KeyboardShortcuts from "~/scenes/KeyboardShortcuts";
 import { createAction } from "~/actions";
 import { NavigationSection, RecentSearchesSection } from "~/actions/sections";
-import env from "~/env";
 import Desktop from "~/utils/Desktop";
 import history from "~/utils/history";
 import isCloudHosted from "~/utils/isCloudHosted";
@@ -212,9 +211,6 @@ export const logout = createAction({
   icon: <LogoutIcon />,
   perform: () => {
     void stores.auth.logout();
-    if (env.OIDC_LOGOUT_URI) {
-      window.location.replace(env.OIDC_LOGOUT_URI);
-    }
   },
 });
