+++ plugins/slack/server/api/hooks.test.ts
@@ -1,6 +1,5 @@
 import randomstring from "randomstring";
 import { IntegrationService } from "@shared/types";
-import env from "@server/env";
 import { IntegrationAuthentication, SearchQuery } from "@server/models";
 import {
   buildDocument,
@@ -9,6 +8,7 @@ import {
   buildUser,
 } from "@server/test/factories";
 import { getTestServer } from "@server/test/support";
+import env from "../env";
 import * as Slack from "../slack";
 
 jest.mock("../slack", () => ({
