+++ plugins/slack/server/slack.ts
@@ -1,7 +1,7 @@
 import querystring from "querystring";
-import env from "@server/env";
 import { InvalidRequestError } from "@server/errors";
 import fetch from "@server/utils/fetch";
+import env from "./env";
 
 const SLACK_API_URL = "https://slack.com/api";
